Image Optimizer

+++ database/migrations/2026_07_17_000001_create_image_optimizer_batches.php
@@ -14,11 +14,11 @@ return new class extends Migration
             $table->foreignId('user_id')->constrained()->cascadeOnDelete();
             $table->string('status', 30)->default('uploading')->index();
             $table->json('settings');
-            $table->unsignedSmallInteger('total_files')->default(0);
-            $table->unsignedSmallInteger('uploaded_files')->default(0);
-            $table->unsignedSmallInteger('processed_files')->default(0);
-            $table->unsignedSmallInteger('completed_files')->default(0);
-            $table->unsignedSmallInteger('failed_files')->default(0);
+            $table->unsignedInteger('total_files')->default(0);
+            $table->unsignedInteger('uploaded_files')->default(0);
+            $table->unsignedInteger('processed_files')->default(0);
+            $table->unsignedInteger('completed_files')->default(0);
+            $table->unsignedInteger('failed_files')->default(0);
             $table->unsignedBigInteger('bytes_total')->default(0);
             $table->unsignedBigInteger('bytes_uploaded')->default(0);
             $table->timestamp('started_at')->nullable();
@@ -36,7 +36,7 @@ return new class extends Migration
                 ->constrained('image_optimizer_batches')
                 ->cascadeOnDelete();
             $table->uuid('uuid')->unique();
-            $table->unsignedSmallInteger('position');
+            $table->unsignedInteger('position');
             $table->string('client_fingerprint', 1000);
             $table->string('relative_path', 1000)->nullable();
             $table->string('original_name');
